factor out test and rename variables in filterDependencies
