+++ packages/modular-scripts/src/__tests__/esmView.test.ts
@@ -14,7 +14,6 @@ import prettier from 'prettier';
 import puppeteer from 'puppeteer';
 
 import getModularRoot from '../utils/getModularRoot';
-import { matchDependencies } from '../utils/filterDependencies';
 import { startApp, DevServer } from './start-app';
 import type { CoreProperties } from '@schemastore/package';
 
@@ -766,6 +765,7 @@ describe('modular-scripts', () => {
       ).toEqual('/public/path/static/js/index-CWZPL7BC.js');
     });
   });
+
   describe('WHEN building a esm-view specifying a PUBLIC_URL and the path is ./', () => {
     beforeAll(async () => {
       await modular('build sample-esm-view', {
@@ -800,82 +800,6 @@ describe('modular-scripts', () => {
           )) as CoreProperties
         ).module,
       ).toEqual('./static/js/index-CWZPL7BC.js');
-    });
-  });
-
-  describe('WHEN we specify an allow / block list', () => {
-    it('THEN expects packages to be blocked / allowed by exact name with blocking winning in case of intersection', () => {
-      const fakePkg = {
-        fake1: '^2.0.1',
-        fake2: '^1.0.1',
-        fake3: '^3.12.5',
-        fake4: '^4.5.36',
-        fake5: '^0.0.1',
-      };
-      const allowList = ['fake1', 'fake2', 'fake3'];
-      const blockList = ['fake2'];
-      expect(
-        matchDependencies({
-          packageDependencies: fakePkg,
-          allowList,
-          blockList,
-        }),
-      ).toEqual({
-        bundled: {
-          fake2: '^1.0.1',
-          fake4: '^4.5.36',
-          fake5: '^0.0.1',
-        },
-        external: {
-          fake1: '^2.0.1',
-          fake3: '^3.12.5',
-        },
-      });
-    });
-    it('THEN expects packages to be blocked / allowed by wildcard', () => {
-      const fakePkg = {
-        'fake1-typea': '^2.0.1',
-        'fake2-typea': '^1.0.1',
-        'fake3-typeb': '^3.12.5',
-        'fake4-typeb': '^4.5.36',
-        'fake5-typec': '^0.0.1',
-      };
-      const allowList = ['**'];
-      const blockList = ['*-typeb'];
-      expect(
-        matchDependencies({
-          packageDependencies: fakePkg,
-          allowList,
-          blockList,
-        }),
-      ).toEqual({
-        bundled: {
-          'fake3-typeb': '^3.12.5',
-          'fake4-typeb': '^4.5.36',
-        },
-        external: {
-          'fake1-typea': '^2.0.1',
-          'fake2-typea': '^1.0.1',
-          'fake5-typec': '^0.0.1',
-        },
-      });
-    });
-    it('THEN expects scoped packages to be allowed by default', () => {
-      const fakePkg = {
-        '@somescope/fake1': '^2.0.1',
-        '@somescope/fake2': '^1.0.1',
-      };
-      expect(
-        matchDependencies({
-          packageDependencies: fakePkg,
-        }),
-      ).toEqual({
-        bundled: {},
-        external: {
-          '@somescope/fake1': '^2.0.1',
-          '@somescope/fake2': '^1.0.1',
-        },
-      });
     });
   });
 });
